telegram_bot/requirements.txt
asyncpg
aiogram
APScheduler
python-decouple
openai
pandas
tiktoken
scipy
